(feature/neeraj-user-activities): routing setup on dashboard
