+++ src/pages/private/Activities/Activities.js
@@ -1,7 +1,11 @@
 import React from "react";
 
 function Activities() {
-  return <h1>Activities</h1>;
+  return (
+    <>
+      <h1>Activities</h1>
+    </>
+  );
 }
 
 export default Activities;
