--- src/pages/private/dashboard/DashboardBody.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-function DashboardBody() {
-  return (
-    <>
-      <h1>DashboardBody</h1>
-      <h1>DashboardBody</h1>
-      <h1>DashboardBody</h1>
-    </>
-  );
-}
-
-export default DashboardBody;
